リファクタリングの再修正
